Refactor Request.init functionality to module
